SOLVED PROBLEM-2
Suppose the Great Frederick Fair wants to update its ticketing software. They need you to write a program to handle the price calculations, using the rules`:
•	The bask price of a ticket is $40.
•	Senior citizens (age s= 65) get a 50% discount.
•	Children under 6 are free (100% discount).
•	For residents of Frederick County, the basic price is $35; the same discounts still apply.
So the individual ticket prices range from $0 to $40.
Your program should request age and county name from the user. The age will be entered as an integer and the county name as a string.
Before calculating the price, confirm that the user's age is valid — not negative and not more than 110. If it is not, give a message and do not do the price calculation. Also, the county name should not be case sensitive — for example, Frederick, frederick, and FREDERICK should all be acceptable.
Your program should then calculate and print out the ticket price, using the appropriate discounts.
Test your program with a variety of ages and counties to be sure you have considered all the conditions. Here are some samples.
Test run #	County	Age
1	Frederick	12
2	Frederick	72
3	Carroll	2
4	Howard	65
5	Washington	0
6	Frederick	5
7	Montgomery	6
8	Carroll	35
9	Frederick	-15
10	Frederick	44
11	Howard	122
12	Cecil	13

Your program should be written with the future in mind. The Great Frederick Fair might need to raise the basic prices or modify the discounts in the future. That means named constants, not hard-coded literals, for the discount rates, age cut-offs, and base price in the calculations.

'These aren't the real prices. The real system is much more complicated. My favorite among the real ones is the Carload Special Tuesday: $60 for everyone legally buckled in a vehicle, buses NOT included.